Fix warnings caused by int64_t <-> size_t conversions (#1381)
This is a fairly large code change, with one main theme:
- Values that are always positive but were stored as `int64_t`, are now converted into `size_t`.

The main driving force for this is the desire to reduce the number of `static_cast`s throughout the code, and allow easier integration with the static library (which overwhelmingly uses `size_t`s for indexes and counts).

This affects, first of all, properties `nrows`, `ncols`, `nkeys` in `DataTable` and `Column` class, but also in some other places. Several functions/methods had their signatures changed to better match the new convention.

In addition, several methods were modified to work with `std::vector`s, instead of plain-C arrays or `arr32_t`s.

+++ c/capi.h
@@ -8,13 +8,14 @@
 #ifndef dt_CAPI_H
 #define dt_CAPI_H
 #include <stdint.h>
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 
-void* datatable_get_column_data(void* dt, int64_t column);
+void* datatable_get_column_data(void* dt, size_t column);
 void  datatable_unpack_slicerowindex(void *dt, int64_t *start, int64_t *step);
 void  datatable_unpack_arrayrowindex(void *dt, void **indices);
