+++ c/frame/_repr_html_.cc
@@ -33,8 +33,8 @@ class HtmlWidget {
       const size_t maxcols = 15;  // TODO: make configurable
       const size_t maxrows = 15;
       dt = dt_;
-      ncols = static_cast<size_t>(dt->ncols);
-      nrows = static_cast<size_t>(dt->nrows);
+      ncols = dt->ncols;
+      nrows = dt->nrows;
       cols0 = ncols <= maxcols ? ncols : maxcols * 2 / 3;
       cols1 = ncols <= maxcols ? 0 : maxcols - cols0;
       rows0 = nrows <= maxrows ? nrows : maxrows * 2 / 3;
